Improve standalone panics
