+++ kernel/standalone/src/main.rs
@@ -32,7 +32,6 @@ mod executor;
 mod hardware;
 mod kernel;
 mod mem_alloc;
-mod panic;
 mod random;
 mod time;
